move common datum code to datum.ts
