+++ packages/studio-base/src/panels/Plot/datum.ts
@@ -1,0 +1,56 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { ScatterDataPoint } from "chart.js";
+
+import { isTime, toSec } from "@foxglove/rostime";
+import { Time } from "@foxglove/studio";
+
+// In addition to the base datum, we also add receiveTime and optionally header stamp to our datums
+// These are used in the csv export.
+export type Datum = ScatterDataPoint & {
+  value: OriginalValue;
+  receiveTime: Time;
+  headerStamp?: Time;
+};
+
+export type OriginalValue = string | bigint | number | boolean | Time;
+
+export function isChartValue(value: unknown): value is OriginalValue {
+  switch (typeof value) {
+    case "bigint":
+    case "boolean":
+    case "number":
+    case "string":
+      return true;
+    case "object":
+      if (isTime(value)) {
+        return true;
+      }
+      return false;
+    default:
+      return false;
+  }
+  return false;
+}
+
+export function getChartValue(value: unknown): number | undefined {
+  switch (typeof value) {
+    case "bigint":
+      return Number(value);
+    case "boolean":
+      return Number(value);
+    case "number":
+      return value;
+    case "object":
+      if (isTime(value)) {
+        return toSec(value);
+      }
+      return undefined;
+    case "string":
+      return +value;
+    default:
+      return undefined;
+  }
+}
